refactor(uat): consolidate free-space measurement into preflight_budget.py (w6)
free_space_gib (the disk-usage measurement primitive) lived in
phases/preflight.py but was consumed by all three disk-policy sites:
preflight's own gate, execute.py's platform-boundary check (imported under
an alias), and the orchestrator's per-cell disk-floor watch (via the
preflight module). Move the measurement into preflight_budget.py -- the
module that already owns the disk-budget estimator it's paired with -- and
have preflight.py re-export it (so its own internal call sites and
test_preflight.py's existing monkeypatches, out of this TODO's scope_limit,
keep working unchanged). execute.py and orchestrator.py now import it
directly from preflight_budget. The three policy sites (thresholds, when to
check, what to do on shortfall) stay distinct; only the measurement is
shared.

Part of uat-execute-path-unification w6. This closes out the TODO's
work-unit sequence (w1-w6).
